Support thinning of context as layers progress away from token embedding.
hyperparameters specify:
a) At what layer to start thinning context.  At the lowest (earliest)
layers the attention across a broad context is critical.  At higher
layers, it starts getting redundant, so we can start thining after
about 2-5 layers

b) How aggressively do we reduce context?  This is done with a
multiplier ratio, which specifies how much smaller the context
supplied to the next layer sohuld be.

c) What kind of thinning is to be done.  We can either discard some
context, or accumulate the context into eldest retained context.
That is, we sum the hidden states that WOULD be discarded, into
the last (eldest hidden state) that we retain.