Correct sum of discards by dividing by the square root of the count
I was incorrectly averaging the discarded hidden states. This meant I
divided by the count of the discarded context.  Better is probably to
divide by the square root of the count, so that the variance of the
states in maintained (roughly).

I also changed the names of the metaparameters controlling the thinning
down of context.